Move @Bean declarations from AnchorPlatformServer to HelperConfig and MetricsConfig for preparation of configuration management refactoring. (#514)
